Some cleanup, better message passing, only done items are deleted when button is pressed.

+++ gui/src/main/resources/serve-from-classpath/todo/dron/observer/todoElementObserver.js
@@ -1,6 +1,7 @@
 /*global require */
 
 define(["lib/templateobserver", "jquery", "underscore","http/json-over-http"],    function (templateService, $, _, jsonOverHttp) {
+    var listener=$("<span>");
     function plugin(element){
         var todoName = element.data("todo-name"),
             todoId = element.data("todo-id"),
@@ -14,15 +15,19 @@ define(["lib/templateobserver", "jquery", "underscore","http/json-over-http"],  
         //if (todoState) element.addClass("todo-done");
         element.append(checkbox).append("<span>"+todoName+"</span>");
 
-        element.on("clearCompleted", function(){
-            jsonOverHttp({uri: '/db/item/'+todoId, method:'DELETE'});
-            element.empty();
+        listener.on("todoClearCompleted", function(){
+            if(element.hasClass("todo-done")){
+                jsonOverHttp({uri: '/db/item/'+todoId, method:'DELETE'});
+                element.empty();
+                element.trigger("todoClearCompleted");
+            }
         });
 
     }
     
     return {
         observer:templateService.observer({method:plugin, elementLocator:"[data-todo-id]"}),
+        trigger:function(event){listener.trigger(event);},
         setJsonOverHttp:function(j){jsonOverHttp = j;}
     };
 });
